chore: Do not execute scripts that render relevant DOM elements only after hydration
