+++ src/app/(nextjs_migration)/(guest)/page.tsx
@@ -18,7 +18,7 @@ export default async function Home() {
 
   return (
     <div data-partial="landing">
-      <script type="module" src="/nextjs_migration/client/js/transitionObserver.js" async></script>
+      <Script type="module" src="/nextjs_migration/client/js/transitionObserver.js" />
       <Script type="module" src="/nextjs_migration/client/js/landing.js" />
       <section className="hero">
         <div>
